requirements.txt
python-telegram-bot
httpx[socks]
python-dotenv
fastapi
uvicorn
jinja2
pycryptodome
protobuf
google-protobuf
aiofiles